Fixes cookie removal on failed authentication, better failed auth messaging
Fix: Previously the cookie was being removed, but as curl was not being closed first, cookie was being recreated on destruct. This is now resolved, by way of ensuring curl_close() is called before unlink.
Update: Added some better feedback for instances where authentication has failed and the account is locked.